[VB-119] adding hideKeyboardWhenTappedAround, fixed layout, refactoring
